requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
sqlmodel==0.0.8
python-jose==3.3.0
httpx==0.24.1
python-dotenv==1.0.0
